requirements.txt
# type checking
mypy

# Cards
svgwrite
Pillow

# psd
psd-tools
